Add content and style to index, add img folder
Created different sections in the html of the website
in order to allow for the user to understand the content
of the website with ease. Added style to each section
such as background colour, visually pleasing positioning,
change in font sizes, etc in order to enhance the users
experience and satisfy them.